+++ src/Graphics/GL/ModelGL.hpp
@@ -45,7 +45,7 @@ namespace anvil
 			 *
 			 * @return  the number of drawn polygons
 			 */
-			int Render(IShader& shader, bool minimal);
+			size_t Render(IShader& shader, bool minimal);
 
 			/**
 			 * @fn	void ModelGL::Update();
