+++ src/Graphics/GL/SkyboxGL.hpp
@@ -52,7 +52,7 @@ namespace anvil
 			 *
 			 * @return returns the number of drawn polygons
 			 */
-			int Render(IShader& shader, bool minimal);
+			size_t Render(IShader& shader, bool minimal);
 
 			/**
 			 * @fn	void Skybox::Update();
